Update: better looking UI and code cleanup
