+++ src/components/PamphletMenu.tsx
@@ -1,6 +1,5 @@
 import React, { useEffect, useState } from "react";
 import {
-  deleteSanghefte,
   getAllPamphlets,
   Pamphlet,
 } from "../util/firestoreFunctions";
@@ -11,26 +10,14 @@ import {
   AccordionItem,
   AccordionPanel,
   Box,
-  Button,
-  Modal,
-  ModalOverlay,
-  ModalContent,
-  ModalHeader,
-  ModalFooter,
-  ModalBody,
-  ModalCloseButton,
-  useDisclosure,
   Flex,
   VStack,
   Container,
   Heading,
   AccordionIcon,
-  Center, Divider
 } from "@chakra-ui/react";
 import { useNavigate } from "react-router-dom";
-import { QRCodeSVG } from "qrcode.react";
-import { SmallAddIcon, AddIcon } from "@chakra-ui/icons";
-import { url } from "inspector";
+import { AddIcon } from "@chakra-ui/icons";
 
 import backgroundImage from "../assets/background_image.jpg"
 
@@ -40,26 +27,6 @@ export const PamphletMenu = () => {
   const userID = localStorage.getItem(localStorageKey);
   const navigate = useNavigate();
 
-  // Magic link
-  const [linkTopamphlet, setLinkToPamphlet] = useState("");
-
-  const { isOpen, onOpen, onClose } = useDisclosure();
-  const [copyLinkBackgroundColor, setCopyLinkBackgroundColor] =
-    useState("white");
-
-  const localStorage_userReferenceKey = "userReference";
-  const userReference = localStorage.getItem(localStorage_userReferenceKey);
-
-  const updateMagicLink = () => {
-    const pamphletTitle = sessionStorage.getItem("currentPamphlet_title");
-    if (userReference !== null && pamphletTitle !== null) {
-      const pamphletMagicLink = encodeURI(
-        "/sing/" + userReference + "/" + pamphletTitle
-      );
-      setLinkToPamphlet("https://sanghefte.no" + pamphletMagicLink);
-    }
-  };
-
   useEffect(() => {
     const fetchPamphlets = async () => {
       setPamphlets([]);
@@ -68,20 +35,10 @@ export const PamphletMenu = () => {
     fetchPamphlets().catch(console.error);
   }, [userID, setPamphlets]);
 
-
-
   const handleClick_createNewPamphlet = () => {
     navigate("/pamphlet");
   };
 
-  function refreshPage() {
-    window.location.reload();
-  }
-
-  const handleClick_addSong = (pamphletTitle: string) => {
-    sessionStorage.setItem("currentPamphlet_title", pamphletTitle);
-    navigate("/newsong");
-  };
   require("qrcode.react");
   return (
     <>
@@ -106,9 +63,9 @@ export const PamphletMenu = () => {
                   <AccordionItem key={pamphlet.id}>
                     <h1>
                       <AccordionButton
-                        onClick={() =>
-                          sessionStorage.setItem("currentPamphlet_title", pamphlet.id)
-                        }
+                        onClick={() => {
+                          sessionStorage.setItem("currentPamphlet_title", pamphlet.id);
+                        }}
                         _expanded={{ bg: "purple.100", border: "0.2px purple", borderRadius: 5 }}
                       >
                         <Box flex={"1"} textAlign={"left"}>{pamphlet.id}</Box>
@@ -117,73 +74,8 @@ export const PamphletMenu = () => {
                     </h1>
                     <AccordionPanel>
                       <PamphletContent pamphletId={pamphlet.id!} />
-                      {/*<br />
-                      <Button onClick={() => handleClick_addSong(pamphlet.id)}>
-                        Legg til sang
-                      </Button>
-                      <Button onClick={() => handleClick_deletePamphlet(pamphlet.id)}>
-                        Slett hefte
-                      </Button>
-                      <Button
-                        onClick={() => {
-                          onOpen();
-                          updateMagicLink();
-                          setCopyLinkBackgroundColor("white");
-                        }}
-                      >
-                        Del hefte
-                      </Button>
-
-                      //Popup-window for Sharing
-                      <Modal isOpen={isOpen} onClose={onClose}>
-                        <ModalOverlay />
-                        <ModalContent>
-                          <ModalHeader>Link til ditt hefte:</ModalHeader>
-                          <ModalCloseButton />
-                          <ModalBody>
-                            <a
-                              href={linkTopamphlet}
-                              target={"_blank"}
-                              rel={"noopener noreferrer"}
-                            >
-                              {linkTopamphlet}
-                            </a>
-                            <QRCodeSVG
-                              value={linkTopamphlet}
-                              size={128}
-                              bgColor={"#ffffff"}
-                              fgColor={"#000000"}
-                              level={"L"}
-                              includeMargin={false}
-                            />
-                          </ModalBody>
-
-                          <ModalFooter>
-                            <Button colorScheme="blue" mr={3} onClick={onClose}>
-                              Lukk
-                            </Button>
-                            <Button
-                              color={
-                                copyLinkBackgroundColor === "white" ? "black" : "white"
-                              }
-                              backgroundColor={copyLinkBackgroundColor}
-                              variant="ghost"
-                              onClick={() => {
-                                navigator.clipboard.writeText(linkTopamphlet);
-                                setCopyLinkBackgroundColor("#4BB543");
-                              }}
-                            >
-                              {copyLinkBackgroundColor === "white"
-                                ? "Kopiér Link"
-                                : "Link Kopiert!"}
-                            </Button>
-                          </ModalFooter>
-                        </ModalContent>
-                      </Modal>*/}
                     </AccordionPanel>
                   </AccordionItem>
-
-
                 ))}
                 <AccordionItem>
                   <AccordionButton onClick={handleClick_createNewPamphlet} marginTop={5} bg={"green.100"} _hover={{bg: "green.200"}}>
